Re-organized helper transformation tests
